feat: initial refactor to enable the Google Vertex AI private provider.
