+++ src/dev_mode/mod.rs
@@ -163,7 +163,7 @@ pub mod test_providers {
         Ok(())
     }
 
-    /// Tests Google provider using gemini-pro
+    /// Tests private Google provider using gemini-pro
     pub(crate) async fn _test_google_provider(
         settings: &Settings,
     ) -> Result<(), Box<dyn std::error::Error>> {
@@ -188,7 +188,7 @@ pub mod test_providers {
                 content: _get_code_str(test_source_file)?,
             }],
         };
-        info!("Prompt data: {:#?}", prompt_data);
+        // info!("Prompt data: {:#?}", prompt_data);
         let result = review_or_summarise(request_type, settings, provider, &prompt_data).await?;
         info!("Result: {:?}", result);
